api/requirements.txt
fastapi==0.121.1
uvicorn==0.24.0
psycopg2-binary==2.9.11
python-dotenv==1.0.0
pydantic==2.12.4
pandas==2.3.3
python-multipart==0.0.6